test: add SubversionRepositoryBuilder
Since they are used in various tests, define them in `test_helper`.
