+++ test/test_helper.rb
@@ -204,3 +204,106 @@ module CommandRunner
     assert(system(*args), "Command failed: #{args.join(' ')}")
   end
 end
+
+module SubversionRepositoryBuilder
+  include CommandRunner
+
+  def create_test_repository(dir)
+    repository_path = File.join(dir, "repository")
+    run_command("svnadmin", "create", repository_path)
+
+    import_path = File.join(dir, "import")
+    FileUtils.mkdir_p(import_path)
+    File.write(File.join(import_path, "a.txt"), "FILE: a.txt\n")
+    File.write(File.join(import_path, "b.txt"), "FILE: b.txt\n")
+
+    repository_url = "file://#{repository_path}"
+    run_command("svn", "import",
+                import_path, "#{repository_url}/dir",
+                "-m", "Add dir/{a.txt,b.txt}")
+    repository_url
+  end
+
+  def build_move_directory_repository(dir)
+    repository_url = create_test_repository(dir)
+    run_command("svn", "move",
+                "#{repository_url}/dir", "#{repository_url}/renamed",
+                "-m", "Move dir to renamed")
+    repository_url
+  end
+
+  def build_2move_directory_repository(dir)
+    repository_url = build_move_directory_repository(dir)
+    run_command("svn", "move",
+                "#{repository_url}/renamed", "#{repository_url}/rerenamed",
+                "-m", "Move dir to rerenamed")
+    repository_url
+  end
+
+  def build_delete_directory_repository(dir)
+    repository_url = create_test_repository(dir)
+    run_command("svn", "rm",
+                "#{repository_url}/dir",
+                "-m", "Delete dir")
+    repository_url
+  end
+
+  def build_copy_directory_repository(dir)
+    repository_url = create_test_repository(dir)
+    run_command("svn", "copy",
+                "#{repository_url}/dir", "#{repository_url}/copied",
+                "-m", "Copy dir to copied")
+    repository_url
+  end
+
+  def build_move_directory_with_deleted_file_repository(dir)
+    repository_url = create_test_repository(dir)
+    # Checkout to perform both `move` and `rm` in the same commit.
+    # (Operations on `repository_url` result in a separate commit for each operation.)
+    work_path = File.join(dir, "work")
+    run_command("svn", "checkout", repository_url, work_path)
+    run_command("svn", "move",
+                File.join(work_path, "dir"), File.join(work_path, "renamed"))
+    run_command("svn", "rm", File.join(work_path, "renamed", "b.txt"))
+    run_command("svn", "commit",
+                "-m", "Move dir to renamed and remove b.txt",
+                work_path)
+    repository_url
+  end
+
+  def build_replace_directory_with_add_repository(dir)
+    repository_url = create_test_repository(dir)
+
+    work_path = File.join(dir, "work")
+    run_command("svn", "checkout", repository_url, work_path)
+    run_command("svn", "rm", File.join(work_path, "dir"))
+    run_command("svn", "mkdir", File.join(work_path, "dir"))
+    File.write(File.join(work_path, "dir", "new.txt"), "FILE: new.txt\n")
+    run_command("svn", "add", File.join(work_path, "dir", "new.txt"))
+    run_command("svn", "commit",
+                "-m", "Replace dir with a new directory",
+                work_path)
+    repository_url
+  end
+
+  def build_replace_directory_with_move_repository(dir)
+    repository_url = create_test_repository(dir)
+
+    import_other_path = File.join(dir, "import_other")
+    FileUtils.mkdir_p(import_other_path)
+    File.write(File.join(import_other_path, "c.txt"), "FILE: c.txt\n")
+    run_command("svn", "import",
+                import_other_path, "#{repository_url}/other",
+                "-m", "Add other/c.txt")
+
+    work_path = File.join(dir, "work")
+    run_command("svn", "checkout", repository_url, work_path)
+    run_command("svn", "rm", File.join(work_path, "dir"))
+    run_command("svn", "move",
+                File.join(work_path, "other"), File.join(work_path, "dir"))
+    run_command("svn", "commit",
+                "-m", "Replace dir with move directory",
+                work_path)
+    repository_url
+  end
+end
